@@ -37,57 +37,56 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.0.0",
-    "@angular/common": "^7.0.0",
-    "@angular/compiler": "^7.0.0",
-    "@angular/core": "^7.0.0",
-    "@angular/forms": "^7.0.0",
-    "@angular/http": "^7.0.0",
-    "@angular/platform-browser": "^7.0.0",
-    "@angular/platform-browser-dynamic": "^7.0.0",
-    "@angular/router": "^7.0.0",
-    "core-js": "^2.5.4",
-    "core.js": "^0.4.2",
-    "rxjs": "^6.3.3",
-    "zone.js": "^0.8.26"
+    "@angular/animations": "^14.2.3",
+    "@angular/common": "^14.2.3",
+    "@angular/compiler": "^14.2.3",
+    "@angular/core": "^14.2.3",
+    "@angular/forms": "^14.2.3",
+    "@angular/http": "^7.2.16",
+    "@angular/platform-browser": "^14.2.3",
+    "@angular/platform-browser-dynamic": "^14.2.3",
+    "@angular/router": "^14.2.3",
+    "core-js": "^3.1.4",
+    "rxjs": "^6.6.0",
+    "zone.js": "^0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.10.2",
-    "@angular-devkit/build-ng-packagr": "^0.10.2",
-    "@angular/cli": "^7.0.2",
-    "@angular/compiler-cli": "^7.0.0",
-    "@angular/language-service": "^7.0.0",
-    "@types/jasmine": "~2.8.6",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "angular-tslint-rules": "^1.5.0",
-    "bundlesize": "^0.17.0",
-    "codelyzer": "^4.5.0",
-    "commitizen": "^2.10.1",
-    "cpy-cli": "^2.0.0",
-    "cz-conventional-changelog": "^2.1.0",
-    "husky": "^0.14.3",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~1.4.2",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "@angular-devkit/build-angular": "^14.2.3",
+    "@angular-devkit/build-ng-packagr": "^0.1002.4",
+    "@angular/cli": "^14.2.3",
+    "@angular/compiler-cli": "^14.2.3",
+    "@angular/language-service": "^14.2.3",
+    "@types/jasmine": "~4.3.0",
+    "@types/jasminewd2": "~2.0.10",
+    "@types/node": "~14.18.31",
+    "angular-tslint-rules": "^1.20.4",
+    "bundlesize": "^0.18.1",
+    "codelyzer": "^6.0.2",
+    "commitizen": "^4.2.5",
+    "cpy-cli": "^4.2.0",
+    "cz-conventional-changelog": "^3.3.0",
+    "husky": "^8.0.1",
+    "jasmine-core": "~4.4.0",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "~6.4.1",
+    "karma-chrome-launcher": "~3.1.1",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
     "karma-mocha-reporter": "^2.2.5",
-    "lint-staged": "^7.2.0",
-    "ng-packagr": "^4.4.0",
-    "prettier": "^1.13.7",
-    "protractor": "~5.3.0",
-    "protractor-browser-logs": "^1.0.351",
+    "lint-staged": "^13.0.3",
+    "ng-packagr": "^14.2.3",
+    "prettier": "^3.0.0-alpha.0",
+    "protractor": "~7.0.0",
+    "protractor-browser-logs": "^1.0.456",
     "protractor-console-plugin": "^0.1.1",
-    "puppeteer": "^0.13.0",
-    "ts-node": "~5.0.1",
-    "tsickle": ">=0.27.3",
-    "tslib": "^1.7.1",
-    "tslint": "^5.11.0",
-    "tslint-config-prettier": "^1.15.0",
-    "typescript": "^3.1.3",
+    "puppeteer": "^18.0.5",
+    "ts-node": "~10.9.1",
+    "tsickle": "0.46.3",
+    "tslib": "^2.4.0",
+    "tslint": "^6.1.3",
+    "tslint-config-prettier": "^1.18.0",
+    "typescript": "~4.6.0",
     "validate-commit-msg": "^2.14.0"
   },
   "lint-staged": {
