@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { ANALYZE_FOR_ENTRY_COMPONENTS, ModuleWithProviders, NgModule, NgModuleFactoryLoader, SystemJsNgModuleLoader, Type } from '@angular/core';
+import { ANALYZE_FOR_ENTRY_COMPONENTS, ModuleWithProviders, NgModule, Type } from '@angular/core';
 import { ROUTES } from '@angular/router';
 // components
 import { LoadableComponent } from './components/loadable.component';
@@ -18,14 +18,14 @@ export class LoadableComponentModule {
    * module root instantiation
    * (used to setup injectable providers used in dynamic component load/render)
    */
-  static forRoot(manifests: Array<LoadableManifest>): ModuleWithProviders {
+  static forRoot(manifests: Array<LoadableManifest>): ModuleWithProviders<LoadableComponentModule> {
     return {
       ngModule: LoadableComponentModule,
       providers: [
         LoadableService,
-        { provide: NgModuleFactoryLoader, useClass: SystemJsNgModuleLoader },
-        // provider for Angular CLI to analyze
-        { provide: ROUTES, useValue: manifests, multi: true },
+        // { provide: NgModuleFactoryLoader, useClass: SystemJsNgModuleLoader },
+
+        { provide: ROUTES, useValue: manifests, multi: true },        // provider for Angular CLI to analyze
         // provider for DynamicComponentLoader to analyze
         { provide: LOADABLE_MANIFESTS, useValue: manifests }
       ]
@@ -36,7 +36,7 @@ export class LoadableComponentModule {
    * feature module import
    * (used to import loadable component helper component to feature modules)
    */
-  static forFeature(): ModuleWithProviders {
+  static forFeature(): ModuleWithProviders<LoadableComponentModule> {
     return {
       ngModule: LoadableComponentModule
     };
@@ -46,13 +46,13 @@ export class LoadableComponentModule {
    * module child instantiation
    * (used by a loadable component module to register themselves as loadable)
    */
-  static forChild(component: Type<any>): ModuleWithProviders {
+  static forChild(component: Type<any>): ModuleWithProviders<LoadableComponentModule> {
     return {
       ngModule: LoadableComponentModule,
       providers: [
         { provide: ANALYZE_FOR_ENTRY_COMPONENTS, useValue: component, multi: true },
         // provider for @angular/router to parse
-        { provide: ROUTES, useValue: [], multi: true },
+        // { provide: ROUTES, useValue: [], multi: true },
         // provider for DynamicComponentLoader to analyze
         { provide: LOADABLE_COMPONENT, useValue: component }
       ]
